backend/requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
passlib[bcrypt]
python-jose
psycopg2-binary
flask
flask-cors
speechrecognition
pyaudio
pocketsphinx
pipwin
python-multipart
python-dotenv